Improve a bit the UI
- Add some cute images
- Add tag version in footer
