+++ .github/workflows/main.yml
@@ -76,6 +76,8 @@ jobs:
         run: |
           chmod u+x "${GITHUB_WORKSPACE}/.github/docker-tag.sh"
           "${GITHUB_WORKSPACE}/.github/docker-tag.sh"
+          JSCVERSION="${{ env.DOCKER_TAG }}"
+          sed -i "s/JSCVERSION/${JSCVERSION}/" "${GITHUB_WORKSPACE}/fe/package.json"
       - name: Set up QEMU
         uses: docker/setup-qemu-action@v1
