+++ fe/.env
@@ -3,3 +3,4 @@ REACT_APP_BACK_END_API_PORT=3001
 REACT_APP_BACK_END_API_SCHEME=http
 REACT_APP_WEBSITE_DESCRIPTION='Cooking tips and tricks, tools to make it all easier.'
 REACT_APP_WEBSITE_TITLE='Cooking for fun'
+REACT_APP_VERSION=$npm_package_version
